requirements.txt
numpy
opencv-python
pillow
photoshop-python-api
mediapipe
requests
torch
torchvision
mmengine>=0.7.1
mmcv>=2.0.0
mmdet>=3.0.0
mmpose>=1.1.0